+++ NEWS.md
@@ -1,6 +1,11 @@
 ucsf-vpn
 ========
 
+## Version (development version)
+
+ * ...
+ 
+
 ## Version 5.8.0 (2024-05-18)
 
 ### Significant changes