+++ NEWS.md
@@ -1,7 +1,7 @@
 ucsf-vpn
 ========
 
-## Version (development version)
+## Version 6.0.0 (2024-05-20)
 
 ### Significant changes